Move @subscribe from ClusterContextManager to ClusterContextManagerBuilder
